rfct: move CLI to dedicated module
