+++ sbb_newspapers_org_image/eynollah.py
@@ -33,7 +33,6 @@ from scipy.signal import find_peaks
 from scipy.ndimage import gaussian_filter1d
 from shapely import geometry
 from lxml import etree as ET
-import click
 from matplotlib import pyplot, transforms
 import matplotlib.patches as mpatches
 import imutils
@@ -11616,24 +11615,3 @@ class eynollah:
 
         print("Job done in: " + str(time.time() - t1))
 
-
-@click.command()
-@click.option("--image", "-i", help="image filename", type=click.Path(exists=True, dir_okay=False))
-@click.option("--out", "-o", help="directory to write output xml data", type=click.Path(exists=True, file_okay=False))
-@click.option("--model", "-m", help="directory of models", type=click.Path(exists=True, file_okay=False))
-@click.option("--save_images", "-si", help="if a directory is given, images in documents will be cropped and saved there", type=click.Path(exists=True, file_okay=False))
-@click.option("--save_layout", "-sl", help="if a directory is given, plot of layout will be saved there", type=click.Path(exists=True, file_okay=False))
-@click.option("--save_deskewed", "-sd", help="if a directory is given, deskewed image will be saved there", type=click.Path(exists=True, file_okay=False))
-@click.option("--save_all", "-sa", help="if a directory is given, all plots needed for documentation will be saved there", type=click.Path(exists=True, file_okay=False))
-@click.option("--allow_enhancement", "-ae", help="if this parameter set to true, this tool would check that input image need resizing and enhancement or not. If so output of resized and enhanced image and corresponding layout data will be written in out directory")
-@click.option("--curved_line", "-cl", help="if this parameter set to true, this tool will try to return contoure of textlines instead of rectabgle bounding box of textline. This should be taken into account that with this option the tool need more time to do process.")
-@click.option("--full_layout", "-fl", help="if this parameter set to true, this tool will try to return all elements of layout.")
-@click.option("--allow_scaling", "-as", help="if this parameter set to true, this tool would check the scale and if needed it will scale it to perform better layout detection")
-@click.option("--headers_off", "-ho", help="if this parameter set to true, this tool would ignore headers role in reading order")
-def main(image, out, model, save_images, save_layout, save_deskewed, save_all, allow_enhancement, curved_line, full_layout, allow_scaling, headers_off):
-    x = eynollah(image, None, out, model, save_images, save_layout, save_deskewed, save_all, allow_enhancement, curved_line, full_layout, allow_scaling, headers_off)
-    x.run()
-
-
-if __name__ == "__main__":
-    main()
